+++ bande/prisma/migrations/20260906080803_lieu_position/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "bande_etiquettes" ADD COLUMN     "latitude" DOUBLE PRECISION,
+ADD COLUMN     "longitude" DOUBLE PRECISION;